- Remove rules that do not apply to apex: await, nameof, as, is, tuple, etc - Update enum definition - Update readme
