+++ test/expressions.tests.ts
@@ -36,101 +36,6 @@ Object newPoint = new Vector(point.x * z, 0);`);
       });
     });
 
-    describe('Anonymous Objects', () => {
-      it('simple - on single line', () => {
-        const input = Input.InMethod(`Object x = new { ID = 42 };`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Object,
-          Token.Identifiers.LocalName('x'),
-          Token.Operators.Assignment,
-          Token.Keywords.Control.New,
-          Token.Punctuation.OpenBrace,
-          Token.Variables.ReadWrite('ID'),
-          Token.Operators.Assignment,
-          Token.Literals.Numeric.Decimal('42'),
-          Token.Punctuation.CloseBrace,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('simple - on multiple lines', () => {
-        const input = Input.InMethod(`
-Object x = new
-{
-    ID = 42
-};`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Object,
-          Token.Identifiers.LocalName('x'),
-          Token.Operators.Assignment,
-          Token.Keywords.Control.New,
-          Token.Punctuation.OpenBrace,
-          Token.Variables.ReadWrite('ID'),
-          Token.Operators.Assignment,
-          Token.Literals.Numeric.Decimal('42'),
-          Token.Punctuation.CloseBrace,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('nested - on single line', () => {
-        const input = Input.InMethod(`Object x = new { y = new { ID = 42 } };`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Object,
-          Token.Identifiers.LocalName('x'),
-          Token.Operators.Assignment,
-          Token.Keywords.Control.New,
-          Token.Punctuation.OpenBrace,
-          Token.Variables.ReadWrite('y'),
-          Token.Operators.Assignment,
-          Token.Keywords.Control.New,
-          Token.Punctuation.OpenBrace,
-          Token.Variables.ReadWrite('ID'),
-          Token.Operators.Assignment,
-          Token.Literals.Numeric.Decimal('42'),
-          Token.Punctuation.CloseBrace,
-          Token.Punctuation.CloseBrace,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('nested - on multiple lines', () => {
-        const input = Input.InMethod(`
-Object x = new
-{
-    y = new
-    {
-        ID = 42
-    }
-};`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Object,
-          Token.Identifiers.LocalName('x'),
-          Token.Operators.Assignment,
-          Token.Keywords.Control.New,
-          Token.Punctuation.OpenBrace,
-          Token.Variables.ReadWrite('y'),
-          Token.Operators.Assignment,
-          Token.Keywords.Control.New,
-          Token.Punctuation.OpenBrace,
-          Token.Variables.ReadWrite('ID'),
-          Token.Operators.Assignment,
-          Token.Literals.Numeric.Decimal('42'),
-          Token.Punctuation.CloseBrace,
-          Token.Punctuation.CloseBrace,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-    });
-
     describe('Arithmetic', () => {
       it('mixed relational and arithmetic operators', () => {
         const input = Input.InMethod(`b = this.i != 1 + (2 - 3);`);
@@ -150,64 +55,6 @@ Object x = new
           Token.Operators.Arithmetic.Subtraction,
           Token.Literals.Numeric.Decimal('3'),
           Token.Punctuation.CloseParen,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-    });
-
-    describe('Await', () => {
-      it('at statement level', () => {
-        const input = Input.InMethod(`await M();`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.Keywords.Await,
-          Token.Identifiers.MethodName('M'),
-          Token.Punctuation.OpenParen,
-          Token.Punctuation.CloseParen,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('in assignment', () => {
-        const input = Input.InMethod(`Object x = await M();`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Object,
-          Token.Identifiers.LocalName('x'),
-          Token.Operators.Assignment,
-          Token.Keywords.Await,
-          Token.Identifiers.MethodName('M'),
-          Token.Punctuation.OpenParen,
-          Token.Punctuation.CloseParen,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('passed as argument', () => {
-        const input = Input.InMethod(`M1(await M2());`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.Identifiers.MethodName('M1'),
-          Token.Punctuation.OpenParen,
-          Token.Keywords.Await,
-          Token.Identifiers.MethodName('M2'),
-          Token.Punctuation.OpenParen,
-          Token.Punctuation.CloseParen,
-          Token.Punctuation.CloseParen,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('statement level (issue #83)', () => {
-        const input = Input.InMethod(`await x;`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.Keywords.Await,
-          Token.Variables.ReadWrite('x'),
           Token.Punctuation.Semicolon
         ]);
       });
@@ -249,28 +96,6 @@ Object x = new
         ]);
       });
 
-      it('cast to tuple type in assignment', () => {
-        const input = Input.InMethod(`Object t = ((Integer x, Integer y))o;`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Object,
-          Token.Identifiers.LocalName('t'),
-          Token.Operators.Assignment,
-          Token.Punctuation.OpenParen,
-          Token.Punctuation.OpenParen,
-          Token.PrimitiveType.Integer,
-          Token.Identifiers.TupleElementName('x'),
-          Token.Punctuation.Comma,
-          Token.PrimitiveType.Integer,
-          Token.Identifiers.TupleElementName('y'),
-          Token.Punctuation.CloseParen,
-          Token.Punctuation.CloseParen,
-          Token.Variables.ReadWrite('o'),
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
       it('passed to invocation', () => {
         const input = Input.InMethod(`M((Integer)42);`);
         const tokens = tokenize(input);
@@ -302,94 +127,6 @@ Object x = new
           Token.Punctuation.CloseParen,
           Token.Literals.Numeric.Decimal('42'),
           Token.Punctuation.CloseParen,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('as cast of identifier', () => {
-        const input = Input.InMethod(`Object x = o as List<Lazy<String>>;`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Object,
-          Token.Identifiers.LocalName('x'),
-          Token.Operators.Assignment,
-          Token.Variables.ReadWrite('o'),
-          Token.Keywords.As,
-          Token.Type('List'),
-          Token.Punctuation.TypeParameters.Begin,
-          Token.Type('Lazy'),
-          Token.Punctuation.TypeParameters.Begin,
-          Token.PrimitiveType.String,
-          Token.Punctuation.TypeParameters.End,
-          Token.Punctuation.TypeParameters.End,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('as cast of invocation', () => {
-        const input = Input.InMethod(`Object x = M() as List<Lazy<String>>;`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Object,
-          Token.Identifiers.LocalName('x'),
-          Token.Operators.Assignment,
-          Token.Identifiers.MethodName('M'),
-          Token.Punctuation.OpenParen,
-          Token.Punctuation.CloseParen,
-          Token.Keywords.As,
-          Token.Type('List'),
-          Token.Punctuation.TypeParameters.Begin,
-          Token.Type('Lazy'),
-          Token.Punctuation.TypeParameters.Begin,
-          Token.PrimitiveType.String,
-          Token.Punctuation.TypeParameters.End,
-          Token.Punctuation.TypeParameters.End,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('is cast of identifier', () => {
-        const input = Input.InMethod(`Object x = o is List<Lazy<String>>;`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Object,
-          Token.Identifiers.LocalName('x'),
-          Token.Operators.Assignment,
-          Token.Variables.ReadWrite('o'),
-          Token.Keywords.Is,
-          Token.Type('List'),
-          Token.Punctuation.TypeParameters.Begin,
-          Token.Type('Lazy'),
-          Token.Punctuation.TypeParameters.Begin,
-          Token.PrimitiveType.String,
-          Token.Punctuation.TypeParameters.End,
-          Token.Punctuation.TypeParameters.End,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('is cast of invocation', () => {
-        const input = Input.InMethod(`Object x = M() is List<Lazy<String>>;`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Object,
-          Token.Identifiers.LocalName('x'),
-          Token.Operators.Assignment,
-          Token.Identifiers.MethodName('M'),
-          Token.Punctuation.OpenParen,
-          Token.Punctuation.CloseParen,
-          Token.Keywords.Is,
-          Token.Type('List'),
-          Token.Punctuation.TypeParameters.Begin,
-          Token.Type('Lazy'),
-          Token.Punctuation.TypeParameters.Begin,
-          Token.PrimitiveType.String,
-          Token.Punctuation.TypeParameters.End,
-          Token.Punctuation.TypeParameters.End,
           Token.Punctuation.Semicolon
         ]);
       });
@@ -670,89 +407,6 @@ Long total = data['bonusGame']['win'].AsLong * data['bonusGame']['betMult'].AsLo
           Token.Punctuation.Semicolon
         ]);
       });
-
-      it('arithmetic expression with multiple element accesses 2 (issue #37)', () => {
-        const input = Input.InMethod(`
-total = data["bonusGame"]["win"].AsLong * data["bonusGame"]["betMult"].AsLong;
-`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.Variables.ReadWrite('total'),
-          Token.Operators.Assignment,
-          Token.Variables.Property('data'),
-          Token.Punctuation.OpenBracket,
-          Token.Punctuation.StringDoubleQuote.Begin,
-          Token.Literals.StringDoubleQuote('bonusGame'),
-          Token.Punctuation.StringDoubleQuote.End,
-          Token.Punctuation.CloseBracket,
-          Token.Punctuation.OpenBracket,
-          Token.Punctuation.StringDoubleQuote.Begin,
-          Token.Literals.StringDoubleQuote('win'),
-          Token.Punctuation.StringDoubleQuote.End,
-          Token.Punctuation.CloseBracket,
-          Token.Punctuation.Accessor,
-          Token.Variables.Property('AsLong'),
-          Token.Operators.Arithmetic.Multiplication,
-          Token.Variables.Property('data'),
-          Token.Punctuation.OpenBracket,
-          Token.Punctuation.StringDoubleQuote.Begin,
-          Token.Literals.StringDoubleQuote('bonusGame'),
-          Token.Punctuation.StringDoubleQuote.End,
-          Token.Punctuation.CloseBracket,
-          Token.Punctuation.OpenBracket,
-          Token.Punctuation.StringDoubleQuote.Begin,
-          Token.Literals.StringDoubleQuote('betMult'),
-          Token.Punctuation.StringDoubleQuote.End,
-          Token.Punctuation.CloseBracket,
-          Token.Punctuation.Accessor,
-          Token.Variables.Property('AsLong'),
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('arithmetic expression with multiple element accesses 3 (issue #37)', () => {
-        const input = Input.InMethod(`
-Long total = (data["bonusGame"]["win"].AsLong) * data["bonusGame"]["betMult"].AsLong;
-`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.PrimitiveType.Long,
-          Token.Identifiers.LocalName('total'),
-          Token.Operators.Assignment,
-          Token.Punctuation.OpenParen,
-          Token.Variables.Property('data'),
-          Token.Punctuation.OpenBracket,
-          Token.Punctuation.StringDoubleQuote.Begin,
-          Token.Literals.StringDoubleQuote('bonusGame'),
-          Token.Punctuation.StringDoubleQuote.End,
-          Token.Punctuation.CloseBracket,
-          Token.Punctuation.OpenBracket,
-          Token.Punctuation.StringDoubleQuote.Begin,
-          Token.Literals.StringDoubleQuote('win'),
-          Token.Punctuation.StringDoubleQuote.End,
-          Token.Punctuation.CloseBracket,
-          Token.Punctuation.Accessor,
-          Token.Variables.Property('AsLong'),
-          Token.Punctuation.CloseParen,
-          Token.Operators.Arithmetic.Multiplication,
-          Token.Variables.Property('data'),
-          Token.Punctuation.OpenBracket,
-          Token.Punctuation.StringDoubleQuote.Begin,
-          Token.Literals.StringDoubleQuote('bonusGame'),
-          Token.Punctuation.StringDoubleQuote.End,
-          Token.Punctuation.CloseBracket,
-          Token.Punctuation.OpenBracket,
-          Token.Punctuation.StringDoubleQuote.Begin,
-          Token.Literals.StringDoubleQuote('betMult'),
-          Token.Punctuation.StringDoubleQuote.End,
-          Token.Punctuation.CloseBracket,
-          Token.Punctuation.Accessor,
-          Token.Variables.Property('AsLong'),
-          Token.Punctuation.Semicolon
-        ]);
-      });
     });
 
     describe('Invocations', () => {
@@ -835,66 +489,6 @@ Long total = (data["bonusGame"]["win"].AsLong) * data["bonusGame"]["betMult"].As
           Token.Identifiers.ParameterName('y'),
           Token.Punctuation.Colon,
           Token.Literals.Numeric.Decimal('23'),
-          Token.Punctuation.CloseParen,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('out argument', () => {
-        const input = Input.InMethod(`M(out x);`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.Identifiers.MethodName('M'),
-          Token.Punctuation.OpenParen,
-          Token.Keywords.Modifiers.Out,
-          Token.Variables.ReadWrite('x'),
-          Token.Punctuation.CloseParen,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('in argument', () => {
-        const input = Input.InMethod(`M(in x);`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.Identifiers.MethodName('M'),
-          Token.Punctuation.OpenParen,
-          Token.Keywords.Modifiers.In,
-          Token.Variables.ReadWrite('x'),
-          Token.Punctuation.CloseParen,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('named out argument', () => {
-        const input = Input.InMethod(`M(x: out y);`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.Identifiers.MethodName('M'),
-          Token.Punctuation.OpenParen,
-          Token.Identifiers.ParameterName('x'),
-          Token.Punctuation.Colon,
-          Token.Keywords.Modifiers.Out,
-          Token.Variables.ReadWrite('y'),
-          Token.Punctuation.CloseParen,
-          Token.Punctuation.Semicolon
-        ]);
-      });
-
-      it('named in argument', () => {
-        const input = Input.InMethod(`M(x: in y);`);
-        const tokens = tokenize(input);
-
-        tokens.should.deep.equal([
-          Token.Identifiers.MethodName('M'),
-          Token.Punctuation.OpenParen,
-          Token.Identifiers.ParameterName('x'),
-          Token.Punctuation.Colon,
-          Token.Keywords.Modifiers.In,
-          Token.Variables.ReadWrite('y'),
           Token.Punctuation.CloseParen,
           Token.Punctuation.Semicolon
         ]);
