+++ test/interface.tests.ts
@@ -25,7 +25,7 @@ describe('Grammar', () => {
     it('interface inheritance', () => {
       const input = `
 interface IFoo { }
-interface IBar : IFoo { }
+interface IBar extends IFoo { }
 `;
 
       const tokens = tokenize(input);
@@ -37,8 +37,8 @@ interface IBar : IFoo { }
         Token.Punctuation.CloseBrace,
         Token.Keywords.Interface,
         Token.Identifiers.InterfaceName('IBar'),
-        Token.Punctuation.Colon,
-        Token.Type('IFoo'),
+        Token.Keywords.Extends,
+        Token.Identifiers.ExtendsName('IFoo'),
         Token.Punctuation.OpenBrace,
         Token.Punctuation.CloseBrace
       ]);
@@ -56,46 +56,6 @@ interface IBar : IFoo { }
         Token.Punctuation.Comma,
         Token.Identifiers.TypeParameterName('T2'),
         Token.Punctuation.TypeParameters.End,
-        Token.Punctuation.OpenBrace,
-        Token.Punctuation.CloseBrace
-      ]);
-    });
-
-    it('generic interface with variance', () => {
-      const input = `interface IFoo<in T1, out T2> { }`;
-      const tokens = tokenize(input);
-
-      tokens.should.deep.equal([
-        Token.Keywords.Interface,
-        Token.Identifiers.InterfaceName('IFoo'),
-        Token.Punctuation.TypeParameters.Begin,
-        Token.Keywords.Modifiers.In,
-        Token.Identifiers.TypeParameterName('T1'),
-        Token.Punctuation.Comma,
-        Token.Keywords.Modifiers.Out,
-        Token.Identifiers.TypeParameterName('T2'),
-        Token.Punctuation.TypeParameters.End,
-        Token.Punctuation.OpenBrace,
-        Token.Punctuation.CloseBrace
-      ]);
-    });
-
-    it('generic interface with constraints', () => {
-      const input = `interface IFoo<T1, T2> where T1 : T2 { }`;
-      const tokens = tokenize(input);
-
-      tokens.should.deep.equal([
-        Token.Keywords.Interface,
-        Token.Identifiers.InterfaceName('IFoo'),
-        Token.Punctuation.TypeParameters.Begin,
-        Token.Identifiers.TypeParameterName('T1'),
-        Token.Punctuation.Comma,
-        Token.Identifiers.TypeParameterName('T2'),
-        Token.Punctuation.TypeParameters.End,
-        Token.Keywords.Where,
-        Token.Type('T1'),
-        Token.Punctuation.Colon,
-        Token.Type('T2'),
         Token.Punctuation.OpenBrace,
         Token.Punctuation.CloseBrace
       ]);
