+++ test/type-name.tests.ts
@@ -34,112 +34,6 @@ describe('Grammar', () => {
       ]);
     });
 
-    it('tuple type - (Integer, Integer)', () => {
-      const input = Input.InClass(`(Integer, Integer) x;`);
-      const tokens = tokenize(input);
-
-      tokens.should.deep.equal([
-        Token.Punctuation.OpenParen,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.Comma,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.CloseParen,
-        Token.Identifiers.FieldName('x'),
-        Token.Punctuation.Semicolon
-      ]);
-    });
-
-    it('tuple type with element names - (Integer i, Integer j)', () => {
-      const input = Input.InClass(`(Integer i, Integer j) x;`);
-      const tokens = tokenize(input);
-
-      tokens.should.deep.equal([
-        Token.Punctuation.OpenParen,
-        Token.PrimitiveType.Integer,
-        Token.Identifiers.TupleElementName('i'),
-        Token.Punctuation.Comma,
-        Token.PrimitiveType.Integer,
-        Token.Identifiers.TupleElementName('j'),
-        Token.Punctuation.CloseParen,
-        Token.Identifiers.FieldName('x'),
-        Token.Punctuation.Semicolon
-      ]);
-    });
-
-    it('nested tuple type - (Integer, (Integer, Integer))', () => {
-      const input = Input.InClass(`(Integer, (Integer, Integer)) x;`);
-      const tokens = tokenize(input);
-
-      tokens.should.deep.equal([
-        Token.Punctuation.OpenParen,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.Comma,
-        Token.Punctuation.OpenParen,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.Comma,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.CloseParen,
-        Token.Punctuation.CloseParen,
-        Token.Identifiers.FieldName('x'),
-        Token.Punctuation.Semicolon
-      ]);
-    });
-
-    it('nested tuple type with element names - (Integer i, (Integer j, Integer k))', () => {
-      const input = Input.InClass(`(Integer i, (Integer j, Integer k)) x;`);
-      const tokens = tokenize(input);
-
-      tokens.should.deep.equal([
-        Token.Punctuation.OpenParen,
-        Token.PrimitiveType.Integer,
-        Token.Identifiers.TupleElementName('i'),
-        Token.Punctuation.Comma,
-        Token.Punctuation.OpenParen,
-        Token.PrimitiveType.Integer,
-        Token.Identifiers.TupleElementName('j'),
-        Token.Punctuation.Comma,
-        Token.PrimitiveType.Integer,
-        Token.Identifiers.TupleElementName('k'),
-        Token.Punctuation.CloseParen,
-        Token.Punctuation.CloseParen,
-        Token.Identifiers.FieldName('x'),
-        Token.Punctuation.Semicolon
-      ]);
-    });
-
-    it('nullable tuple type - (Integer, Integer)?', () => {
-      const input = Input.InClass(`(Integer, Integer)? x;`);
-      const tokens = tokenize(input);
-
-      tokens.should.deep.equal([
-        Token.Punctuation.OpenParen,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.Comma,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.CloseParen,
-        Token.Punctuation.QuestionMark,
-        Token.Identifiers.FieldName('x'),
-        Token.Punctuation.Semicolon
-      ]);
-    });
-
-    it('array tuple type - (Integer, Integer)[]', () => {
-      const input = Input.InClass(`(Integer, Integer)[] x;`);
-      const tokens = tokenize(input);
-
-      tokens.should.deep.equal([
-        Token.Punctuation.OpenParen,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.Comma,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.CloseParen,
-        Token.Punctuation.OpenBracket,
-        Token.Punctuation.CloseBracket,
-        Token.Identifiers.FieldName('x'),
-        Token.Punctuation.Semicolon
-      ]);
-    });
-
     it('generic type - List<Integer>', () => {
       const input = Input.InClass(`List<Integer> x;`);
       const tokens = tokenize(input);
@@ -148,44 +42,6 @@ describe('Grammar', () => {
         Token.Type('List'),
         Token.Punctuation.TypeParameters.Begin,
         Token.PrimitiveType.Integer,
-        Token.Punctuation.TypeParameters.End,
-        Token.Identifiers.FieldName('x'),
-        Token.Punctuation.Semicolon
-      ]);
-    });
-
-    it('generic type with tuple - List<(Integer, Integer)>', () => {
-      const input = Input.InClass(`List<(Integer, Integer)> x;`);
-      const tokens = tokenize(input);
-
-      tokens.should.deep.equal([
-        Token.Type('List'),
-        Token.Punctuation.TypeParameters.Begin,
-        Token.Punctuation.OpenParen,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.Comma,
-        Token.PrimitiveType.Integer,
-        Token.Punctuation.CloseParen,
-        Token.Punctuation.TypeParameters.End,
-        Token.Identifiers.FieldName('x'),
-        Token.Punctuation.Semicolon
-      ]);
-    });
-
-    it('generic type with tuple with element names - List<(Integer i, Integer j)>', () => {
-      const input = Input.InClass(`List<(Integer i, Integer j)> x;`);
-      const tokens = tokenize(input);
-
-      tokens.should.deep.equal([
-        Token.Type('List'),
-        Token.Punctuation.TypeParameters.Begin,
-        Token.Punctuation.OpenParen,
-        Token.PrimitiveType.Integer,
-        Token.Identifiers.TupleElementName('i'),
-        Token.Punctuation.Comma,
-        Token.PrimitiveType.Integer,
-        Token.Identifiers.TupleElementName('j'),
-        Token.Punctuation.CloseParen,
         Token.Punctuation.TypeParameters.End,
         Token.Identifiers.FieldName('x'),
         Token.Punctuation.Semicolon
